Move file meli-auth.js from '/lib' to '/config' since it's more config-logic related than an external lib.
