+++ lib/meli-client.js
@@ -1,7 +1,7 @@
 const Promise = require('bluebird')
 const req = require('request-promise')
 const Account = require('../model/account')
-const {refresh} = require('./meli-auth')
+const {refresh} = require('../config/meli-auth')
 
 const GET_ORDERS_URL = `https://api.mercadolibre.com/orders/search`
